简单的实现Tree显示分类信息，还不完善

+++ BBSS/src/main/java/com/booksales/model/Book.java
@@ -21,7 +21,7 @@ public class Book {
 
 	private String repertory;
 
-	private Integer price;
+	private Double price;
 
 	private Integer stock;
 
@@ -109,11 +109,12 @@ public class Book {
 		this.repertory = repertory == null ? null : repertory.trim();
 	}
 
-	public Integer getPrice() {
+
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(Integer price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
